donation system and event bus

+++ API/Commands/Donate/DonateCommand.cs
@@ -6,14 +6,14 @@ namespace Infrastructure.Models.Commands
 {
     public class DonateCommand : ICommand<DonationConfirmation>
     {
-        public DonateCommand(int userId, ulong amount, string currency)
+        public DonateCommand(ulong userId, ulong amount, string currency)
         {
             UserId = userId;
             Amount = amount;
             Currency = currency;
         }
 
-        public int UserId { get; }
+        public ulong UserId { get; }
         public ulong Amount { get; }
         public string Currency { get; }
     }
